Remove overflow hidden from content_region
This was causing cropping on some of the content. I think it was a hack to get
something else work.

Letters shouldn't be overrunning, but until we have negitive letter spacing, I
suppose it will be there in the MVP at least, and the best thing is to no do a
hard overflow hidden.