Link between pages
This is mostly for expirmentation reasons, but is also simply a different way of
doing things. If we are linking between stories, we are going to want to use a
story slug as the anchor id, and that is probably the better more symantic way
to go over using page references.