Fix issue with after element for rendering broken text
This is based on emperical observation. Because I dropped the fallback fonts in
the previous commit, it caused the trick I'm using to get full span on the last
line of text in a broken paragraph to stop working. Adding in serif so it has
something to fallback on, probably because the glyphs I'm providing don't work
at the size speicifed or something for some reason. It doesn't really matter,
this seems to work.