Simplify so that masthead is on first page
This is sort of a change in direction, but the goal is to keep the masthead
within the boundries of the first page, rather than on top. This helps with
layout and keeping a consistant size for each page. I use the template name as a
key to look up a masthead in the masthead templates, and I use the wrap_with
helper to target things as at a default masthead template that includes some
general inclusions. Probably overkill, and should probably call it the base
masthead, rather than default, as there could be a default mast head.